Post Model - one like per user functionality

+++ controllers/postControllers.js
@@ -1,3 +1,4 @@
+const { Comment } = require("../models/Comment");
 const { Post, validatePost } = require("../models/Post");
 const asyncHandler = require("express-async-handler");
 
@@ -43,15 +44,24 @@ const getAllPosts = asyncHandler(async (req, res) => {
 
   const totalPosts = await Post.countDocuments(searchQuery)
 
-  const posts = await Post.find(searchQuery)
+  let posts = await Post.find(searchQuery)
     .sort(sortQuery)
     .limit(limit)
     // POPULATING AUTHOR FIELD WITH USERNAME
     .populate("author", "_id username profile")
     .populate("categories", "name _id");
+
+    // Populate comments count for each post
+    posts = await Promise.all(
+      posts.map(async (post) => {
+      const commentsCount = await Comment.countDocuments({ postId: post._id });
+      return { ...post.toObject(), commentsCount };
+    })
+);    
   
   res.status(200).send({
     posts,
+    // postsWithComments,
     totalPosts,
   });
 });
@@ -60,10 +70,15 @@ const getAllPosts = asyncHandler(async (req, res) => {
 const getSinglePost = asyncHandler(async (req, res) => {
   const { id } = req.params;
   const post = await Post.findById(id)
-    .populate("author", "username _id")
-    .populate("categories", "name _id");
+    .populate("author", "_id profile username")
+    .populate("categories", "_id name");
+
+  const commentsCount = await Comment.countDocuments({postId: post._id})
   if (!post) return res.status(400).send({ message: "No post found!" });
-  res.status(200).send(post);
+  res.status(200).send({
+    post,
+    commentsCount
+  });
 });
 
 // POST REQUEST
@@ -104,8 +119,16 @@ const updatePost = asyncHandler(async (req, res) => {
 // DELETE REQUEST
 const deletePost = asyncHandler(async (req, res) => {
   const {id} = req.params
-  const post = await Post.findByIdAndDelete(id)
-  if (!post) return res.status(400).send({message: "No post found!"}) 
+  const post = await Post.findById(id)
+  if (!post) return res.status(400).send({message: "Post not found!"}) 
+
+  // ONLY OWNER & ADMIN ALLOWED
+  if (req.user.id !== post.author && req.user.roles.indexOf("Admin") === -1){
+    return res.status(403).send({message: "You are not allowed to delete this post!"})
+  }
+
+  await Post.findByIdAndDelete(id)
+
   res.status(200).send({message: "Post deleted successfully!"})
 })
 
@@ -114,21 +137,18 @@ const likePost = asyncHandler(async (req, res) => {
   const post = await Post.findById(id);
   if (!post) return res.status(404).json({ message: 'Post not found' });
   
-  post.likes += 1;
-  await post.save();
-
-  res.status(200).json({ message: "You've appreciated this post! ❤️" });
-})
-
-const unlikePost = asyncHandler(async (req, res) => {
-  const {id} = req.params
-  const post = await Post.findById(id);
-  if (!post) return res.status(404).json({ message: 'Post not found' });
-  
-  post.likes -= 1;
-  await post.save();
-
-  res.status(200).json({ message: "Like removed! Your interaction matters! 🙌" });
+  const userLikeIndex = post.likedBy.indexOf(req.user.id);
+  if (userLikeIndex === -1) {
+    post.likes += 1;
+    post.likedBy.push(req.user.id);
+    await post.save();
+    res.status(200).send({ message: "You've appreciated this post! ❤️" });
+  } else {
+    post.likes -= 1;
+    post.likedBy.splice(userLikeIndex, 1);
+    await post.save();
+    res.status(200).send({ message: "Like removed! Your interaction matters! 🙌" });
+  }
 })
 
 const viewPost = asyncHandler(async (req, res) => {
@@ -153,6 +173,5 @@ module.exports = {
   updatePost,
   deletePost,
   likePost,
-  unlikePost,
   viewPost,
 };
